Revert "move docker msmarco test to its own files"
This reverts commit 077d4409e7296eafb6246ac7c62eab4b2fd3228c.
